nineteenth task

+++ src/com/epam/task12/Main.java
@@ -8,7 +8,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
 
-        double a = in.nextDouble(), result = 0;
+        long a = in.nextLong(), result = 0;
 
         for (int i = 1; i < a; i++) {
             result += i;
